+++ openradar/tasks.py
@@ -27,7 +27,10 @@ os.environ['PYTHONPATH'] = ':'.join(sys.path)
 
 # Configure celery
 app = celery.Celery()
-app.conf.update(BROKER_URL=config.CELERY_BROKER_URL)
+app.conf.update(
+    BROKER_URL=config.CELERY_BROKER_URL,
+    CELERYD_TASK_TIME_LIMIT=300,
+)
 
 
 @celery.task